README: Add swagger-js-codegen.queries for the multi-task
'swagger-js-codegen' has been a multi-task since 6173c732 (First
commit, 2014-06-15), and multi-tasks need named sub-properties to
iterate over [1].  I don't see much point to both having a multi-task
and an array for 'apis'.  Perhaps we should have:

  'swagger-js-codegen': {
    batch: {
      options: {
        api: {
          swagger: 'swagger/_batch.json',
          moduleName: 'Batch'
        },
        dest: 'lib'
      },
    },
    queries: {
      options: {
        api: {
          swagger: 'swagger/_queries.json',
          moduleName: 'Queries'
        },
        dest: 'lib'
      },
    },
  }

which would allow both the complete:

  $ grunt swagger-js-codegen

and the per-API:

  $ grunt swagger-js-codegen:queries

but that's food for another commit.

[1]: http://gruntjs.com/api/grunt.task#grunt.task.registermultitask